+++ src/api/domain/Title.ts
@@ -1,0 +1,15 @@
+export class Title {
+  private value: string;
+
+  get Value(): string {
+    return this.value;
+  }
+
+  constructor(value: string) {
+    this.value = value;
+  }
+
+  public equals(other: Title): boolean {
+    return this.Value === other.Value;
+  }
+}
