+++ src/api/domain/Todo.test.ts
@@ -1,4 +1,7 @@
-import { CompletedAt, CreatedAt, DueDate, Todo } from "./Todo";
+import { CompletedAt } from "./CompletedAt";
+import { CreatedAt } from "./CreatedAt";
+import { DueDate } from "./DueDate";
+import { Todo } from "./Todo";
 
 describe("Todo", () => {
   it("やることを生成する", () => {
