+++ src/api/domain/Todo.ts
@@ -1,3 +1,8 @@
+import { CompletedAt } from "./CompletedAt";
+import { CreatedAt } from "./CreatedAt";
+import { DueDate } from "./DueDate";
+import { Title } from "./Title";
+
 export class Todo {
   private title: Title;
   private completed: boolean;
@@ -74,76 +79,3 @@ export class Todo {
     );
   }
 }
-
-export class Title {
-  private value: string;
-
-  get Value(): string {
-    return this.value;
-  }
-
-  constructor(value: string) {
-    this.value = value;
-  }
-
-  public equals(other: Title): boolean {
-    return this.Value === other.Value;
-  }
-}
-
-export class CreatedAt {
-  private value: Date;
-
-  get Value(): Date {
-    return this.value;
-  }
-
-  constructor(value: Date) {
-    this.value = value;
-  }
-
-  public equals(createdAt: CreatedAt): boolean {
-    return this.value.getTime() === createdAt.Value.getTime();
-  }
-}
-
-export class CompletedAt {
-  private value: Date | null;
-
-  get Value(): Date | null {
-    return this.value;
-  }
-
-  constructor(value: Date | null) {
-    this.value = value;
-  }
-
-  public equals(completedAt: CompletedAt): boolean {
-    return this.value?.getTime() === completedAt.Value?.getTime();
-  }
-}
-
-export class DueDate {
-  private value: Date | null;
-
-  get Value(): Date | null {
-    return this.value;
-  }
-
-  constructor(value: Date | null) {
-    this.value = value;
-  }
-
-  public equals(dueDate: DueDate): boolean {
-    return this.value?.getTime() === dueDate.Value?.getTime();
-  }
-
-  public overDue(): boolean {
-    const due = this.Value;
-    const today = new Date();
-    if (due === null) {
-      return false;
-    }
-    return today.getTime() > due.getTime();
-  }
-}
