+++ src/api/index.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import cors from "cors";
-import { Todo } from "./Todo";
+import { Todo } from "./domain/Todo";
 const app = express();
 app.use(cors({ origin: true, credentials: true }));
 app.use(
